requirements.txt
transformers
torch
diffusers
opencv-python
Pillow
gradio
requests
pyyaml
newspaper3k
google-genai
